tests iniciales/requirements.txt
numpy
pandas
pyyaml
scikit-learn
pyod
jinja2